# Conflicts:
#	src/main/java/org/craftercms/studio/impl/v1/service/deployment/PublishingManagerImpl.java

+++ src/main/java/org/craftercms/studio/impl/v1/service/deployment/PublishingManagerImpl.java
@@ -375,19 +375,18 @@ public class PublishingManagerImpl implements PublishingManager {
                 
                 if (oldPath.endsWith("/" + DmConstants.INDEX_FILE)) {
                     String fullPath = contentService.expandRelativeSitePath(site, oldPath.replace("/" + DmConstants.INDEX_FILE, ""));
-                    
-                    RepositoryItem[] children = contentRepository.getContentChildren(site, oldPath.replace("/" + DmConstants.INDEX_FILE, ""));
-                    
-                    if (children.length < 2) {
-                        deployer.deleteFile(site, oldPath.replace("/" + DmConstants.INDEX_FILE, ""));
+                    if (contentService.contentExists(site, oldPath.replace("/" + DmConstants.INDEX_FILE, "")) {
+                        RepositoryItem[] children = contentRepository.getContentChildren(site, oldPath.replace("/" + DmConstants.INDEX_FILE, ""));
+
+                        if (children.length < 2) {
+                            deployer.deleteFile(site, oldPath.replace("/" + DmConstants.INDEX_FILE, ""));
+                        } else {
+                            hasRenamedChildren = true;
+                        }
+
                     }
-                    else {
-                        hasRenamedChildren = true;
-                    }
-                    
-                    
                     if (!hasRenamedChildren) {
-                        contentService.deleteContent(site, oldPath.replace("/" + DmConstants.INDEX_FILE, ""), false, user);
+                        deleteFolder(site, oldPath.replace("/" + DmConstants.INDEX_FILE, ""), user, deployer);
                     }
                 }
 
@@ -402,12 +401,14 @@ public class PublishingManagerImpl implements PublishingManager {
             if (item.getPath().endsWith("/" + DmConstants.INDEX_FILE)) {
                 String fullPath = contentService.expandRelativeSitePath(site, path.replace("/" + DmConstants.INDEX_FILE, ""));
                 RepositoryItem[] children = contentRepository.getContentChildren(site, path.replace("/" + DmConstants.INDEX_FILE, ""));
+                if (contentService.contentExists(site, path.replace("/" + DmConstants.INDEX_FILE, ""))) {
+                    RepositoryItem[] children = contentRepository.getContentChildren(site, path.replace("/" + DmConstants.INDEX_FILE, ""));
 
-                if (children.length < 2) {
-                    deployer.deleteFile(site, path.replace("/" + DmConstants.INDEX_FILE, ""));
-                }
-                else {
-                    haschildren = true;
+                    if (children.length < 2) {
+                        deployer.deleteFile(site, path.replace("/" + DmConstants.INDEX_FILE, ""));
+                    } else {
+                        haschildren = true;
+                    }
                 }
             }
 
@@ -415,7 +416,7 @@ public class PublishingManagerImpl implements PublishingManager {
                 contentService.deleteContent(site, path, user);
                 
                 if (!haschildren) {
-                    contentService.deleteContent(site, path.replace("/" + DmConstants.INDEX_FILE, ""), false, user);
+                    deleteFolder(site, path.replace("/" + DmConstants.INDEX_FILE, ""), user, deployer);
                 }
             }
         }
@@ -464,7 +465,7 @@ public class PublishingManagerImpl implements PublishingManager {
                         
                         
                         if (!hasRenamedChildren) {
-                            contentService.deleteContent(site, oldPath.replace("/" + DmConstants.INDEX_FILE, ""), false, user);
+                            deleteFolder(site, oldPath.replace("/" + DmConstants.INDEX_FILE, ""), user, deployer);
                         }
                     }
                     
@@ -523,9 +524,18 @@ public class PublishingManagerImpl implements PublishingManager {
         }
     }
 
-    @Override
-    public void setLockBehaviourEnabled(boolean enabled) {
-        //contentRepository.setLockBehaviourEnabled(enabled);
+    private void deleteFolder(String site, String path, String user, Deployer deployer) {
+        String fullPath = contentService.expandRelativeSitePath(site, path);
+        if (contentService.contentExists(fullPath)) {
+            RepositoryItem[] children = contentRepository.getContentChildren(fullPath);
+
+            if (children.length < 1) {
+                contentService.deleteContent(site, path, false, user);
+                deployer.deleteFile(site, path);
+                String parentPath = ContentUtils.getParentUrl(path);
+                deleteFolder(site, parentPath, user, deployer);
+            }
+        }
     }
 
     @Override
